fix clash due to redefinition of macro in monolithic build mode: this definition is now identical to the one present in another extension source file.

+++ ext/rtree/rtree.c
@@ -438,7 +438,7 @@ struct RtreeMatchArg {
 #ifndef SQLITE_AMALGAMATION
 # if defined(SQLITE_COVERAGE_TEST) || defined(SQLITE_DEBUG)
     unsigned int sqlite3RtreeTestcase = 0;
-#   define testcase(X)  if( X ){ sqlite3RtreeTestcase += __LINE__; }
+#   define testcase(X)  if( X ){ sqlite3RtreeTestcase += (unsigned)__LINE__; }
 # else
 #   define testcase(X)
 # endif
